+++ .gitignore
@@ -15,6 +15,8 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.idea
+*.iml
 
 npm-debug.log*
 yarn-debug.log*
