+++ .npmrc
@@ -1,0 +1,4 @@
+proxy=http://155.55.60.117:8088/
+registry=http://registry.npmjs.org/
+no-proxy=*.adeo.no
+strict-ssl=false
